correcting to respect magic casting topology fixing visual bug in the README markdown to prevent the getting started guide from appearing as a solid block of text instead of a hierarchy of bullet points in MarkdownPad 2
